Replace GSON.
