+++ openchrom/plugins/net.openchrom.installer.ui/META-INF/MANIFEST.MF
@@ -34,7 +34,6 @@ Require-Bundle: org.eclipse.core.runtime,
  org.eclipse.equinox.p2.ui;bundle-version="2.8.100",
  org.eclipse.chemclipse.progress;bundle-version="0.9.0",
  net.openchrom.feature.branding;bundle-version="1.6.13",
- com.google.gson;bundle-version="2.13.2",
  org.eclipse.e4.core.contexts;bundle-version="[1.13.0,2.0.0)"
 Service-Component: OSGI-INF/net.openchrom.installer.ui.OpenchromP2Policy.xml
 Bundle-RequiredExecutionEnvironment: JavaSE-21
