+++ openchrom/plugins/net.openchrom.installer.ui/src/net/openchrom/installer/ui/wizards/PluginDiscoveryWizardMainPage.java
@@ -92,8 +92,8 @@ import net.openchrom.installer.model.DiscoveryCategory;
 import net.openchrom.installer.model.IPluginDescriptor;
 import net.openchrom.installer.model.PluginDescriptorKind;
 import net.openchrom.installer.model.PluginDiscovery;
+import net.openchrom.installer.model.SetupDefinition;
 import net.openchrom.installer.preferences.PreferenceSupplier;
-import net.openchrom.installer.ui.model.SetupDefinition;
 import net.openchrom.installer.ui.swt.OverviewToolTip;
 import net.openchrom.installer.util.DiscoveryCategoryComparator;
 import net.openchrom.installer.util.DiscoveryConnectorComparator;
