+++ openchrom/plugins/net.openchrom.installer/META-INF/MANIFEST.MF
@@ -15,7 +15,9 @@ Require-Bundle: org.eclipse.core.runtime,
  org.eclipse.e4.ui.workbench;bundle-version="0.10.100",
  org.eclipse.chemclipse.rcp.ui.icons;bundle-version="0.8.0",
  org.eclipse.chemclipse.logging;bundle-version="0.8.0",
- org.eclipse.chemclipse.support;bundle-version="0.9.0"
+ org.eclipse.chemclipse.support;bundle-version="0.9.0",
+ com.fasterxml.jackson.core.jackson-databind;bundle-version="2.21.0",
+ com.fasterxml.jackson.core.jackson-core;bundle-version="2.21.1"
 Export-Package: net.openchrom.installer.model,
  net.openchrom.installer.preferences,
  net.openchrom.installer.util
